requirements.txt
Django==3.2.9
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
